feat: enhance args management

+++ internal/helpers/shortner.go
@@ -33,7 +33,7 @@ var client = &http.Client{
 // Shortner takes a URL string and an optional keepAliveFor duration in seconds,
 // sends a POST request to the endpoint with the URL as the target,
 // and returns the shortened URL string and any error encountered.
-func Shortner(url string, keepAliveFor *int64) (string, error) {
+func Shortner(url string, keepAliveFor *int) (string, error) {
 
 	requestBody, err := json.Marshal(map[string]string{
 		"target":       url,
